[ALFREDOPS-864] bug fix for integration tests
