+++ integration-tests/src/main/amp/config/alfresco/extension/subsystems/HealthProcessor/default/default/example-extensions-context.xml
@@ -4,10 +4,6 @@
         xsi:schemaLocation="http://www.springframework.org/schema/beans
 		http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="eu.xenit.alfresco.healthprocessor.example.ExampleHealthProcessorPlugin"
-            class="eu.xenit.alfresco.healthprocessor.example.ExampleHealthProcessorPlugin">
-        <constructor-arg name="nodeService" ref="NodeService"/>
-        <property name="enabled" value="${eu.xenit.alfresco.healthprocessor.plugin.example.enabled}"/>
-    </bean>
+
 
 </beans>
