+++ integration-tests/src/main/amp/config/alfresco/module/alfresco-health-processor-example/module-context.xml
@@ -7,45 +7,65 @@
     <!--
         Proxy required to make webscript available in main Spring context
     -->
-    <bean id="webscript.eu.xenit.alfresco.healthprocessor.example.extensions.info.get"
-            class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
+
+    <bean id="eu.xenit.alfresco.healthprocessor.plugins.solr.endpoint.SearchEndpointSelector"
+          class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
         <property name="sourceApplicationContextFactory" ref="HealthProcessor"/>
-        <property name="sourceBeanName" value="ExampleExtensionsGetInfoWebScript"/>
+        <property name="sourceBeanName" value="eu.xenit.alfresco.healthprocessor.plugins.solr.endpoint.SearchEndpointSelector"/>
         <property name="interfaces">
             <list>
-                <value>org.springframework.extensions.webscripts.WebScript</value>
+                <value>eu.xenit.alfresco.healthprocessor.plugins.solr.endpoint.SearchEndpointSelector</value>
             </list>
         </property>
     </bean>
 
+    <bean id="eu.xenit.alfresco.healthprocessor.fixer.solr.SolrMissingNodeFixerPlugin"
+          class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
+        <property name="sourceApplicationContextFactory" ref="HealthProcessor"/>
+        <property name="sourceBeanName" value="eu.xenit.alfresco.healthprocessor.fixer.solr.SolrMissingNodeFixerPlugin"/>
+        <property name="interfaces">
+            <list>
+                <value>eu.xenit.alfresco.healthprocessor.fixer.solr.SolrMissingNodeFixerPlugin</value>
+            </list>
+        </property>
+    </bean>
+
+    <bean id="nodeFinder"
+          class="eu.xenit.alfresco.healthprocessor.solr.NodeFinder">
+        <constructor-arg name="nodeService" ref="NodeService" />
+        <constructor-arg name="searchService" ref="SearchService" />
+        <constructor-arg name="namespaceService" ref="NamespaceService" />
+    </bean>
+
+    <bean id="eu.xenit.alfresco.healthprocessor.example.ExampleHealthProcessorPlugin"
+          class="eu.xenit.alfresco.healthprocessor.example.ExampleHealthProcessorPlugin">
+        <constructor-arg name="nodeService" ref="NodeService"/>
+        <property name="enabled" value="true"/>
+    </bean>
+
+    <bean id="webscript.eu.xenit.alfresco.healthprocessor.example.extensions.info.get"
+          class="eu.xenit.alfresco.healthprocessor.example.ExampleExtensionsGetInfoWebScript"
+          parent="webscript">
+        <constructor-arg name="plugin" ref="eu.xenit.alfresco.healthprocessor.example.ExampleHealthProcessorPlugin"/>
+    </bean>
+
     <bean id="webscript.eu.xenit.alfresco.healthprocessor.solr.purge.get"
-            class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
-        <property name="sourceApplicationContextFactory" ref="HealthProcessor"/>
-        <property name="sourceBeanName" value="solr.PurgeWebScript"/>
-        <property name="interfaces">
-            <list>
-                <value>org.springframework.extensions.webscripts.WebScript</value>
-            </list>
-        </property>
+          class="eu.xenit.alfresco.healthprocessor.solr.SolrPurgeNodeWebScript"
+          parent="webscript">
+        <constructor-arg name="nodeFinder" ref="nodeFinder" />
+        <constructor-arg name="endpointSelector" ref="eu.xenit.alfresco.healthprocessor.plugins.solr.endpoint.SearchEndpointSelector" />
+        <constructor-arg name="globalProperties" ref="global-properties" />
     </bean>
+
     <bean id="webscript.eu.xenit.alfresco.healthprocessor.solr.check.get"
-            class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
-        <property name="sourceApplicationContextFactory" ref="HealthProcessor"/>
-        <property name="sourceBeanName" value="solr.CheckWebScript"/>
-        <property name="interfaces">
-            <list>
-                <value>org.springframework.extensions.webscripts.WebScript</value>
-            </list>
-        </property>
+          class="eu.xenit.alfresco.healthprocessor.solr.SolrCheckNodeWebScript"
+          parent="webscript">
+        <constructor-arg name="nodeFinder" ref="nodeFinder" />
     </bean>
+
     <bean id="webscript.eu.xenit.alfresco.healthprocessor.solr.configure.get"
-            class="org.alfresco.repo.management.subsystems.SubsystemProxyFactory">
-        <property name="sourceApplicationContextFactory" ref="HealthProcessor"/>
-        <property name="sourceBeanName" value="solr.ConfigureWebScript"/>
-        <property name="interfaces">
-            <list>
-                <value>org.springframework.extensions.webscripts.WebScript</value>
-            </list>
-        </property>
+          class="eu.xenit.alfresco.healthprocessor.solr.SolrConfigureIndexNodeFixerPlugin"
+          parent="webscript" autowire="constructor">
     </bean>
+
 </beans>
